Adding support for request bodies for PUT and POST methods, and nested parameters for PUT / POST method request bodies.
Functionality description:
This patch adds support for documenting parameters of a request body. As
values are filled into the content parameters, the method's text area is
filled in with what the request body should be. The text area is what
will be submitted as the request body, and can be modified directly if a
user wants to.

Implementation details:
Content parameters use nested lists instead of tables. It was easier to
figure out how to style nested lists compared to styling nested tables.
Styling was done using Sass and Compass, making use of the updated
style.scss.

In app.js, the 'requestBody' variable has been replaced by the 'content'
variable.